feat: implement authorization check
