+++ project/src/store/api-action.ts
@@ -1,7 +1,8 @@
-import { APIRoute } from '../const';
-import { loadOffers, setIsDataLoading } from './reducers/app-data/app-data';
-import { ThunkActionResult } from '../types/action';
 import { adaptDataToClient } from './adapter';
+import { APIRoute, AuthorizationStatus } from '../const';
+import { loadOffers, setIsDataLoading } from './reducers/app-data/app-data';
+import { setAuthorizationStatus } from './reducers/user-data/user-data';
+import { ThunkActionResult } from '../types/action';
 import { toast } from 'react-toastify';
 
 export const fetchOffersAction = (): ThunkActionResult =>
@@ -20,4 +21,10 @@ export const fetchOffersAction = (): ThunkActionResult =>
     finally {
       dispatch(setIsDataLoading(false));
     }
+  };
+
+export const checkAuthAction = (): ThunkActionResult =>
+  async (dispatch, _getState, api) => {
+    await api.get(APIRoute.Login);
+    dispatch(setAuthorizationStatus(AuthorizationStatus.Auth));
   };
